Breakout_AI_Task/requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
# asttokens==0.7.0
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.32.0
azure-identity==1.19.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
gspread 
oauth2client 
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
comm==0.1.4
cryptography==43.0.3
dataclasses-json==0.6.7
datasets==3.1.0
debugpy==1.8.1
decorator==5.1.1
flask-cors
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.9.0
faiss-gpu==1.7.2
filelock==3.16.1
Flask==3.0.3
frozenlist==1.5.0
fsspec==2024.9.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
lxml[html_clean]
aiohttp
lxml
requests_html
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.23.5
idna==3.10
importlib_metadata==7.1.0
inquirerpy==0.3.4
InstructorEmbedding==1.0.1
interchange==2021.0.4
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
llama-index-core==0.11.22
llama-index-embeddings-azure-openai==0.2.5
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-instructor==0.2.1
llama-index-embeddings-openai==0.2.5
llama-index-graph-stores-neo4j==0.3.5
llama-index-llms-azure-openai==0.2.2
llama-index-llms-groq==0.2.0
llama-index-llms-huggingface==0.3.5
llama-index-llms-huggingface-api==0.2.0
llama-index-llms-openai==0.2.16
llama-index-llms-openai-like==0.2.0
llama-index-readers-file==0.2.2
llama-index-vector-stores-faiss==0.2.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.6
mdurl==0.1.2
minijinja==2.2.0
monotonic==1.6
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
neo4j==5.26.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.54.3
packaging==24.0
pandas==2.2.3
pansi==2024.11.0
parso==0.8.3
pexpect==4.9.0
pfzy==0.3.4
pickleshare==0.7.5
pillow==11.0.0
platformdirs==4.2.0
portalocker==2.10.1
prompt_toolkit==3.0.43
propcache==0.2.0
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.2
py2neo==2021.2.4
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
PyJWT==2.9.0
pypdf==4.3.1
python-dateutil==2.9.0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.2
regex==2024.11.6
reportlab==4.2.5
requests>=2.32.2
rich==13.4.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
scraper-api==0.1.0
scraperapi-sdk==1.5.3
sentence-transformers==2.7.0
serpapi==0.1.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.2
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
text-generation==0.7.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.5.1
tornado==6.4
tqdm==4.67.0
traitlets==5.14.2
transformers==4.46.2
triton==3.1.0
typing-inspect==0.9.0
typing_extensions>=4.11,<5
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.16.0
xxhash==3.5.0
yarl==1.17.1
zipp==3.17.0